Make use of the debug CLI flag for request response status and error printing
